ENH:restore support for compiling without C99 complex support
This is needed for MSVC compatiblity.

This requires using defines in C (DEF in Cython) to control whether the complex valued
routines are compiled.  On the Python side, _have_c99_complex is used to control
whether complex data can be passed directly to the Cython DWT and SWT routines or if
the real and imaginary parts must be processed separately.